fix(portable-text-editor): remove invalid return value
We should not concider the state of isPatching here to determine if we are saving the operation or not
as it leads to errors when there are multiple users in the same doc.

This was maybe true at some point, but it is now invalid and leads to errors in the common history.

The intention of this return is now taken care of with better scoping elsewhere in the code (withouthSaving)
where non-undo-redo operations are created (as with remote patching)