fix(portable-text-editor): show editable when invalid value
Always show the editable also when there is an invalid value.

This is making a better experience when choosing to ignore the validation error for a PT text node.

The invalid value is now removed from the greater value when displayed in the editor,
making it possible to still edit everything else if you choose to. This was not
previously possible, but became possible with the introduction of the useSyncValue hook.